% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CATEGORY_DERIVED_DATA}
\alias{CATEGORY_DERIVED_DATA}
\title{Category Derived Data}
\format{
An object of class \code{character} of length 1.
}
\usage{
CATEGORY_DERIVED_DATA
}
\description{
Category Derived Data
}
\keyword{datasets}
