% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CATEGORY_FOLLOW_UP}
\alias{CATEGORY_FOLLOW_UP}
\title{Category Follow Up}
\format{
An object of class \code{character} of length 1.
}
\usage{
CATEGORY_FOLLOW_UP
}
\description{
Category Follow Up
}
\keyword{datasets}
