% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CATEGORY_INTERVENTION}
\alias{CATEGORY_INTERVENTION}
\title{Category Intervention}
\format{
An object of class \code{character} of length 1.
}
\usage{
CATEGORY_INTERVENTION
}
\description{
Category Intervention
}
\keyword{datasets}
