% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CATEGORY_NICU_DISCHARGE}
\alias{CATEGORY_NICU_DISCHARGE}
\title{Category NICU Discharge}
\format{
An object of class \code{character} of length 1.
}
\usage{
CATEGORY_NICU_DISCHARGE
}
\description{
Category NICU Discharge
}
\keyword{datasets}
