% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CATEGORY_PERIOD_MAP}
\alias{CATEGORY_PERIOD_MAP}
\title{Category Period mapping}
\format{
An object of class \code{list} of length 6.
}
\usage{
CATEGORY_PERIOD_MAP
}
\description{
Category Period mapping
}
\keyword{datasets}
