% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CATEGORY_POST_INTERVENTION}
\alias{CATEGORY_POST_INTERVENTION}
\title{Category Post-intervention}
\format{
An object of class \code{character} of length 1.
}
\usage{
CATEGORY_POST_INTERVENTION
}
\description{
Category Post-intervention
}
\keyword{datasets}
