% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CATEGORY_PRE_INTERVENTION}
\alias{CATEGORY_PRE_INTERVENTION}
\title{Category Pre-intervention}
\format{
An object of class \code{character} of length 1.
}
\usage{
CATEGORY_PRE_INTERVENTION
}
\description{
Category Pre-intervention
}
\keyword{datasets}
