% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{corr}
\alias{corr}
\title{Compute Pearson's correlation coefficient analysis between x-columns and y-columns}
\usage{
corr(df, x_columns, y_columns)
}
\arguments{
\item{df}{Dataframe to process}

\item{x_columns}{x-columns for Pearson's correlation coefficient analysis}

\item{y_columns}{y-columns for Pearson's correlation coefficient analysis}
}
\value{
Dataframe with ycolumns as (r, R^2, p-value, valid, valid-x, valid-y, valid-percent)
}
\description{
Compute Pearson's correlation coefficient analysis between x-columns and y-columns
}
