% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DATA_DICTIONARY_EXCEL}
\alias{DATA_DICTIONARY_EXCEL}
\title{Dictionary spreadsheet}
\format{
An object of class \code{character} of length 1.
}
\usage{
DATA_DICTIONARY_EXCEL
}
\description{
Dictionary spreadsheet
}
\keyword{datasets}
