% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DESCRIPTION}
\alias{DESCRIPTION}
\title{Variable description}
\format{
An object of class \code{character} of length 1.
}
\usage{
DESCRIPTION
}
\description{
Variable description
}
\keyword{datasets}
