% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{FILENAME_INFOS}
\alias{FILENAME_INFOS}
\title{Filename info}
\format{
An object of class \code{list} of length 72.
}
\usage{
FILENAME_INFOS
}
\description{
Filename info
}
\keyword{datasets}
