% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten_column_tuple}
\alias{flatten_column_tuple}
\title{flatten column tuple}
\usage{
flatten_column_tuple(column)
}
\description{
Flatten column of the flattened csv to the tuple (prefix, var-name, postfix)
In R, the columns become X[prefix].[var-name].[postfix]
}
