% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{get_inv_mapping_value}
\alias{get_inv_mapping_value}
\title{Get inverse-value from inv_mapping}
\usage{
get_inv_mapping_value(the_type, value)
}
\arguments{
\item{the_type}{the type of the value (sheet_name)}

\item{value}{the value}
}
\value{
inverse value (or NA if not available)
}
\description{
Get inverse-value from inv_mapping
}
