% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{init_inv_mapping}
\alias{init_inv_mapping}
\title{Initialize inverse mapping from standardized variables to the corresponding rank.}
\usage{
init_inv_mapping(filename, column)
}
\arguments{
\item{filename}{dictionary filename}

\item{column}{the column name for inv mapping.}
}
\description{
Initialize inverse mapping from standardized variables to the corresponding rank.
}
