% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dict.R
\name{load_data_dict}
\alias{load_data_dict}
\title{Load Data Dictionary with filling the category column}
\usage{
load_data_dict(filename, sheet_name)
}
\arguments{
\item{filename}{data dictionary filename}

\item{sheet_name}{sheet name to load}
}
\description{
Load Data Dictionary with filling the category column
}
