% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{P_VALUE_THRESHOLD}
\alias{P_VALUE_THRESHOLD}
\title{Significance level / threshold for p-value}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
P_VALUE_THRESHOLD
}
\description{
Significance level / threshold for p-value
}
\keyword{datasets}
