% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SHEET_DERIVED_DATA}
\alias{SHEET_DERIVED_DATA}
\title{Reserved tab: derived data}
\format{
An object of class \code{character} of length 1.
}
\usage{
SHEET_DERIVED_DATA
}
\description{
Reserved tab: derived data
}
\keyword{datasets}
