% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SHEET_FOLLOW_UP}
\alias{SHEET_FOLLOW_UP}
\title{Reserved tab: follow-up}
\format{
An object of class \code{character} of length 1.
}
\usage{
SHEET_FOLLOW_UP
}
\description{
Reserved tab: follow-up
}
\keyword{datasets}
