% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SHEET_MAIN}
\alias{SHEET_MAIN}
\title{Reserved tab: main}
\format{
An object of class \code{character} of length 1.
}
\usage{
SHEET_MAIN
}
\description{
Reserved tab: main
}
\keyword{datasets}
