% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SHEET_NAMING_CONVENTION}
\alias{SHEET_NAMING_CONVENTION}
\title{Reserved tab: naming convention}
\format{
An object of class \code{character} of length 1.
}
\usage{
SHEET_NAMING_CONVENTION
}
\description{
Reserved tab: naming convention
}
\keyword{datasets}
