% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SHEET_RESERVED_TABS}
\alias{SHEET_RESERVED_TABS}
\title{Reserved tabs in the spreadsheet}
\format{
An object of class \code{character} of length 5.
}
\usage{
SHEET_RESERVED_TABS
}
\description{
Reserved tabs in the spreadsheet
}
\keyword{datasets}
