% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SHEET_VARIABLES}
\alias{SHEET_VARIABLES}
\title{Reserved tabs in the spreadsheet
Reserved tab: Variables}
\format{
An object of class \code{character} of length 1.
}
\usage{
SHEET_VARIABLES
}
\description{
Reserved tabs in the spreadsheet
Reserved tab: Variables
}
\keyword{datasets}
