% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VAR_NAME}
\alias{VAR_NAME}
\title{Standardized variable names dictionary}
\format{
An object of class \code{character} of length 1.
}
\usage{
VAR_NAME
}
\description{
Standardized variable names dictionary
}
\keyword{datasets}
